refactor: Merge utils, typeGuards and helpers
