+++ packages/client/src/pages/Signup/SignUp.tsx
@@ -2,20 +2,21 @@ import { FC, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { Navigate } from 'react-router-dom';
 
-import { authApi } from '../../api/Auth/Auth';
-import { AppLink } from '../../components/AppLink/AppLink';
-import { FormButton } from '../../components/FormButton/FormButton';
-import { Header } from '../../components/Header';
-import { Input } from '../../components/Input/Input';
 import {
   validateEmail,
   validateLogin,
   validateNames,
   validatePassword,
   validatePhone,
-} from '../../helpers/authFormValidation';
+} from '@utils/authFormValidation';
+import { isNetworkError } from '@utils/isNetworkError';
+
+import { authApi } from '../../api/Auth/Auth';
+import { AppLink } from '../../components/AppLink/AppLink';
+import { FormButton } from '../../components/FormButton/FormButton';
+import { Header } from '../../components/Header';
+import { Input } from '../../components/Input/Input';
 import { RoutePaths } from '../../providers/Router/AppRouter/constants';
-import { isNetworkError } from '../../typeGuards/isNetworkError';
 import { TFormValues } from './types';
 
 export const SignUp: FC = () => {
