+++ packages/client/src/utils/validationHelpers.ts
